Add ZARC "ruling" tutorial
ie. difference between monster type and archetype.
Ruling isn't the best word here, but I'm not sure what else to put. I intend it to be consistent with the Xyz Material thing, so if you change this, change that and future ones too. (Also the AI name in the puzzles.)